requirements.txt
mcstatus==6.*
discord.py==1.*
python-a2s==1.*
